+++ app/src/test/kotlin/me/xizzhu/android/joshua/translations/TranslationPresenterTest.kt
@@ -87,8 +87,10 @@ class TranslationPresenterTest : BaseUnitTest() {
             downloadedTranslationsChannel.send(listOf(MockContents.kjvDownloadedTranslationInfo))
 
             val expected = listOf(
+                    TitleItem("English", true),
                     TranslationItem(MockContents.kjvDownloadedTranslationInfo, true, translationPresenter::onTranslationClicked, translationPresenter::onTranslationLongClicked),
                     TitleItem("", false),
+                    TitleItem("Chinese", true),
                     TranslationItem(MockContents.cuvTranslationInfo, false, translationPresenter::onTranslationClicked, translationPresenter::onTranslationLongClicked)
             )
             verify(translationView, times(1)).onTranslationsUpdated(expected)